about to remove the networking script because of library troubles; will readd later when it is next desired
